Extracted `stripMarkdown()`
